Added initial helm chart
This commit adds initial support for installing the KubeDirector
Deployment via Helm.  Values are provided to override the number of
replicas, the CRD names, image name/tag, and the service account names.
This chart is configured to pull from bluek8s/kubedirector:unstable from
the master branch and should be updated in other release branches.

Related to #63