Add node_id support for clusters.
Some apps like zookeeper need to determine which node will
be the first to initialize, second and third etc. Such apps cannot
depend on any other naming schema for determining that info as
it may change when the cluster is expanded.

To help those apps, we have a monotonically increasing number
assigned as an ID to each pod that is part of the cluster. The
bdmacro API provides access to the ID.