Make sure modules haven't changed since the last commit.
I think that go modules are probably the way forward for
dependency management. This commit adds two new make
commands - modules and verify-modules - that will ensure
we're ready to uptake this. The modules command should
automatically regenerate the go.mod and go.sum files
(which manage what modules we depend on, similar to
Gopkg.lock and Gopkg.toml) and the verify-modules
command tests that our modules are all up-to-date. The
verify-modules command is mainly meant to be used by
Travis-CI, but it might work on user's machines as well.

Link to an interesting post about go's modules:
https://github.com/golang/go/wiki/Modules