don't allow app delete/modify while in use
Addresses issue #7.

This change adds another map in the handler state, which associates a "cluster key" (namespace+name) to the app ID used by that cluster. We're going to use that in-memory object to track when a KubeDirectorApp is referenced by any KubeDirectorCluster, and prevent app delete/modify in that situation. The functions that reference the map have a bit more commentary about why that data structure is used.

Note that if you want to try out this branch, you'll probably want to set your Local.mk to deploy from my bluek8s/kubedirector:joel-dev image. Because this changes the webhook properties it's not easy to get it running correctly just by using "make redeploy" after deploying from the normal image.

----------

File-by-file changes:

### reconciler/types.go

Adding the above-mentioned map to the handler state.

### reconciler/handler.go

Initializing the map.

### reconciler/util.go

Functions for reading/writing the map.

### reconciler/cluster.go

If we're discovering existing clusters on KD startup (for example during a redeploy), go ahead and add them to the map.

### validator/util.go

Adding "DELETE" to the set of REST ops that our webhook will pay attention to.

### validator/cluster.go

Add and remove clusters from the map as appropriate.

Also, when processing an update, we don't need to do a K8s API GET to obtain the pre-update resource. We can just get it out of the OldObject in the admission request.

Minor improvement: Don't validate anything (other than the protections for the Status) if the spec is not changing. Some comments in the code about that.

### validator/app.go

On delete or update, check to see if the app is in use; fail admission if so.