move app-in-use tracking to reconciler
The problem w/ doing it in the admission handler is that K8s may still reject the request after we allow it. So in the admission handler we don't actually know for sure that the CR will get created or deleted.

Doing the tracking in the reconciler is more solid, but it does open us up to a tiny race window where an app CR could be deleted right after we admit a new cluster CR but before the reconciliation handler runs for it. In that case the cluster CR will be created but the handler will return an error and the CR will never get reconciled. I'll file a separate issue for further thinking about that.