prevent crash when scaling role
issue #51

Rather than just checking to see if roleStatus is nil, we can check to see if there are zero ready members. This similarly handles the zero-members case but will also skip non-zero-members roles that are entirely made up of non-ready members. In that case we also don't need to bother finding the setup package URL.