fixing up misc things (mainly app-in-use) after merge
Minor things:
* "make compile" target needs to depend on the generated code being up-to-date.
* Example CR for non-persistent CentOS cluster had a bad name.

The major things though are related to the app-in-use marking...

Previously the validator had been nailing down the app namespace by writing it into the status. With the status being a subresource now we shouldn't do that (and probably just shouldn't regardless; let's leave the status to be controlled only by the reconciler). Also it's a little bad that we don't let the user make the app namespace decision.

So we'll manage it with an optional property in the spec that specifies whether the app is coming from the local namespace or the system (KD) namespace. If left unspecified this will behave same as before, where we look first in local namespace then in system namespace -- and the property will then be populated (by the validator) with local/system value depending on where the app was found.

This affects the implementation of FindApp/GetApp as even FindApp will honor that namespace setting if it was user-specified.

Another change I made was in the format of the map used to store the app references, so apptypes.go has been churned. Previously we had to handle the fact that we were doing the "remove" without having the cluster CR available, so the only thing we knew was the cluster CR/namespace and we had to iterate through the map to find the app it usedd. Now however we always know the name+namespace of both the app and cluster CRs for both add and remove, so we can do more efficient lookups. The map is now indexed by app name+namespace and the value is a list of the clusters using the app.

I modified handleNewCluster slightly so that this is always the spot where "ensure" is done for new clusters, whether or not this KD process has seen them before. I also added more comments.