some cleanup
Some cleanup that happened during work to bring in a controller for another resource type.

Mostly this is formatting, fixing comments, and making some generic var/function names more specific.

For service update, I removed a redundant log message and fixed the arguments to another couple of args.

I also switched the argument order for GetSecret to be consistent with the other getters.