camelCase our CR property names
This is the style guide for Kubernetes resource properties, and "kubectl describe" works best with camelCased names. So let's bite the bullet.

One bullet not bitten here however is the naming scheme for the config data seen inside the container by config CLI. Changing those property names, although possibly nice for consistency with some parts of the CRs, would require also changing config CLI and app startscripts. I'm not sure if we want to do that, and even if we do, let's not bundle it into this change here.