allow restricting the set of mountable secrets
KubeDirector can mount any secret in a KubeDirectorCluster's namespace if it is given a valid secret resource name. There's some more discussion of the security implications on the wiki, and we have upcoming plans to support secrets with encrypted contents where the encryption key can be specified/managed without exposing it. That being said...

We also should allow the KD site configuration to restrict the set of secrets that KubeDirector can mount, by specifying a required name prefix. This allows an admin to be fundamentally sure about which secrets KD will never allow to be mounted/exposed.

By default there will not be any required prefix, but if the global config specifies a prefix then that will be the first thing KD checks at the time of validating a secret mount (it is not retroactive for existing mounts). If the prefix match fails, then that failure is the only thing we will report; we will not check for & report on whether the secret exists in the namespace.