omnibus of improvements from work on other forks
Significant improvements to the CRDs to make them check more stuff and to get rid of errors reported by K8s on the created CRD resources. These changes do require using the "nullable" attribute, which in turn requires using K8s 1.14 or later. Also we need to be more careful about when we specify "omitempty" for our type fields.

Use the API rbac.authorization.k8s.io/v1 instead of rbac.authorization.k8s.io/v1beta1. This non-beta version has been available for a long time and beta may go away.

Improvements to the GKE usage notes.

Some significant changes to how we handle finalizers in our CR reconciliation handlers. The previous logic relied on the first creation of the finalizer happening before any status stanza existed (so that status protection would not be triggered by the write). This is OK for now but it is kind of a time bomb; in some work on another operator we found instances where we might need to affect the status stanza during handleFinalizers. Best way to make this possible is to NOT to do an object write inside of RemoveFinalizer or EnsureFinalizer. Instead those functions can just modify the CR in memory, and finalizers can be written back in the deferfunc just like we do with status changes.

Misc log message and comment improvements.

Generic-izing executor.ownerReferences so that owner references can be created to any kind of object.

Adding the ability to specify labels to add to pods in your KubeDirectorCluster spec.