robustification
We should never let an invalid CR into etcd. This will totally screw up any further reconciliation attempts. So, K8s should reject an attempt to create/modify (or in some cases even an attempt to delete) one of our CRs if our admission control hook is down. To ensure this, I've changed the failure policy to "Fail" for some of our webhooks.

In the event that someone really needs to delete a CR and KD is down and can't be resurrected, then they will have to edit or remove the webhook.

I've also added a liveness probe. If our validation server is down (probably because the entire operator process is dead), K8s will restart the container after 30 seconds of dead-ness. This should also help to solve the issue where the operator binary isn't running inside the container after a host reboot -- I would still like to be able to root-cause that, but I haven't made progress on that.