use trimpath in "make compile"
This removes $GOPATH from the beginning of the filepaths built into the object, and e.g. shown in stacktraces.

The operator-sdk build is supposed to do this as well, although with our version of the operator SDK you need to backport this fix to your own local copy before building/installing it: https://github.com/operator-framework/operator-sdk/pull/1535