prevent cluster spec change if previous spec change unprocessed
This is really unlikely to happen, but if it does happen (perhaps because handler taking really long?) it would be bad.

One badness is that validateRoleChanges might get the wrong answer. Also our upcoming changes to better deal with pod error states. Basically anything here in the validator that relies on looking at the cluster status stanza is made reliable by enforcing this invariant.