track the configured container ID
This will be used to determine if the container got restarted (and therefore we might have something we need to do to it). The restart counter is not reliable.

We need to get the container ID from executor.ReadFile, so for symmetry and possible future use I'll make the other similar guest-op functions also return container ID.