refresh config on member restart
If the container ID changes on a ready or config-error member, boot it back to create pending state.

Various bits of app setup tweaked to acknowledge that it could be a restart. Also note that if we are restarting after a config error, or if there's no persistent storage, we want to completely rerun the setup.