handled config-interrupted-by-restart with persistent storage
If there's an existing config status file with no final status in it yet, we need to know if maybe it was created by a previous incarnation of this container (and therefore the setup script is not actually running anymore and needs to be restarted). To do this we'll stash the container ID in that status file.