don't try to update status if deleting
With all our new kdcluster status it can be the case that some things in status have changed even as we're removing our finalizer, and because we have essentially cleaned up our internal object state at this point the status update will be rejected by the validator.

Making the same change for config CR too although I don't think it is currently vulnerable.

(Obviously we need to factor out this status-mgmt code and share it.)