don't include status stanza in Update for finalizers
Anything specified here is a NOP anyway since status is now a subresource, but it can trigger validation failure in the case that a) we did make a change to the status but b) we are ignoring the status updated because the object is being deleted.

It should be OK to incur the cost of DeepCopy here since finalizer update is rare.