set SpecGenerationToProcess earlier
If we get an error in syncMemberServices (causing handler exit), we lose the necessary update to SpecGenerationToProcess.

Easiest way to trigger this is to run KD in an environment like GKE using LoadBalancer services that can take a while to delete. Create a cluster with a large number of role members, then edit the cluster to shrink that role to its minimum, then immediately edit the cluster again to size the role up again. Because the old services are still being deleted and we want to re-use the same service names, reconciliation will fail until the old services finish deleting.

We should set SpecGenerationToProcess as soon as we know we should. Right after syncClusterRoles should be good, since the "state" from that function is what we would be worried about losing an an error bailout. We don't need to worry about currentHash and cr.Status.LastConnectionHash since we'll get the same values for those on the next handler (unless currentHash legitimately changes).

With this change the above test scenario works, i.e. reconciliation proceeds after the old services go away.

Previously merged to master branch as part of PR #316; cherry-picking (modified) to 0.4.2.