allow PUT to an in-use kdapp if nothing is being changed
closes issue #319

It's not super-trivial to see whether an incoming spec changes anything, because of the way we handle global defaults and apply them to roles. So the validateRoles function here is changed to not only generate patches to modify the resulting document, but also apply the same changes to the in-memory appCR so we can use it for comparisons if this is a PUT.

Then admitAppCR does the comparison. There's a big comment added to admitAppCR so more about that below.

==========

This change also fixes a minor bug in our defaults handling. The pattern is that we apply the global default to the roles (if necessary) and then remove the default from the spec. However in the case of defaultConfigPackage, if it was explicitly set to null, we were not removing it.

Among our own example apps, this only happened in our "vanilla OS" apps (for CentOS etc.) and is functionally harmless, but I needed to fix this behavior so that everything worked uniformly and I could do the necessary comparisons.

That does mean that any previously-created kdapps with a null defaultConfigPackage would not be able to take advantage of this change. That exception might kind of mess with upgrades so I made the additional change in admitAppCR to ignore any difference in defaultConfigPackage value. These defaults don't directly affect how kdclusters behave in any case; they only affect the roles (which we ARE checking for differences).