if kdcluster already has finalizer, treat it as a restored CR
(Note that any incoming status stanza will be dumped because we are using a status subresource.)

Implements the following behaviors in the validator:

- if operation is create and already has finalizer, set being-restored label and skip validation; accept
- if operation is update and has being-restored label, reject any spec change
- if operation is delete and has being-restored label, reject unless allow-force-delete is set

Because we're doing more early-exits from the cluster CR validate function BUT may still need to do patches in those cases, I moved the patch-apply code into a defer func so it always runs on function exit. This defer func also handles errors, and changing the allowed result to true if there are no errors.

The intended use for these labels:

The "being restored" label will, in an upcoming commit, be checked by the reconciliation code. Reconciliation will be skipped if this label is set. The user can manually delete this label to force reconciliation to restart; or, a later commit will handle checking to see if all necessary resources have been restored so that it can un-set this label automatically.

A user cannot normally delete a kdcluster that is in the "being restored" state. This is because other resources may be restored which depend on the kdcluster as their "owner", for proper cleanup when the kdcluster is deleted. However if the user is double-sure that they really do want to delete the kdcluster, and they don't want to remove the "being restored" label (because that would restart reconciliation), they can manually set another label to force the delete to be allowed.