don't rely on finalizer to detect "restored" CR
Backup solution might strip out finalizers. Use an explicit annotation instead. As a bonus we can use the value of this annotation to memo-ize whether or not a status backup exists, to let the user know whether waiting is likely to resolve anything.

Note that if you change the global config setting -- so that status backups are created or deleted for all kdclusters when they reconcile -- this means that every kdcluster will update its annotation as well as doing a no-op "status update" to create/delete the backup. This generates a bang-bang rapid sequence of two changes to the kdcluster, so that we are guaranteed to generate the harmless-but-messy "stale UID" message in the reconciler when the handlers then run for those two changes (since the first will be stale). I think that's fine, probably not worth the effort to avoid the no-op status update in that situation.