secretKeys documentation and fixes
The main point of this PR is to add a doc about "secret keys" that can be referenced from the CRD wiki pages.

I've also bundled a fix for an upgrade case: make sure that the master key default is generated if there was a pre-existing kdconfig when KD was upgraded.

I think it's quite likely that we'll want to include some immediate fixes for issue #512 in here as well, along with the necessary doc updates to reflect that.