implement wait-for-connection (configurable)
Implement the ability to wait for connections before resuming reconcile, and put a boolean in kd-global-config to choose whether to do this. By default we will now wait.

Updated the docs for this, and also fixed an error in the docs where the "awaiting" flags were described as initially-false rather than as initially-true.