make sure we get the correct configcli.tgz on build
Don't trust that the existing configcli.tgz is from the tag indicated by configcli_version.

Let's download configcli using its original archive file name (which includes the version) and just symbolically link that to configcli.tgz (so the Dockerfile can use an unchanging filename). The downloaded file's name can be used to check if it's the version we need.