"legacy" links to support old app images/scripts
The confgcli scripts have been moved from /usr/bin to /usr/local/bin, so that we don't need to put any /usr stuff on the PV. However some old images and scripts may have the /usr/bin paths hardcoded.

We could require updates to app images if the app wants to enable the new layout. However we could also bend over backwards to support those old apps and make symlinks into /usr/bin. This means that if an app is otherwise OK with the effects of enabling the new layout, it doesn't have to change its image or setup package... just the kdapp resource, which is a much more lightweight change.

The "gotcha" of course is that since /usr is not (by default) persisted, those symlinks will go away if the pod container is restarted. So we re-create them before any invocations of the startscript by KD. This means that the symlinks are not guaranteed to be there except for KD-invoked startscript runs (i.e. script triggers for lifecycle events).

If an app is not OK with that limitation for some reason, then yeah it will need to change its own scripts to use the new configcli locations under /usr/local/bin.