better persistDirs union-ing
With the new "tighter" default persist dirs it could be possible that the app requests some dir that is a parent of a default dir. So when we're constructing the final list of mounts, we can't just (as previously) check if each app dirs is a subdir/dup of any default dir. Instead we need to check all the dirs for being subdir/dup of any other.