support custom size for /dev/shm
Allow specifying a "sharedMemory" quantity in the kdcluster role spec, to specify the max size of the memory backing /dev/shm. If not set it defaults to 64MB.

Using this feature requires the SizeMemoryBackedVolumes feature in the K8s cluster. That is a feature flag that is false by default in K8s 1.20/1.21, and true by default in K8s 1.22+. It hasn't graduated out of being a feature flag as of K8s 1.24.