allow startscript to open the /etc/guestconfig perms
This allows for an intermediate step of app development if an app needs to adopt the new persistent dirs layout but would be broken by the new more restrictive perms on /etc/guestconfig. This lets the startscript be explicit about the need/desire to share configmeta.json with other user accounts, without KD interfering.

The functional change in KD is that we will only set the 700 perms on /etc/guestconfig when first creating that directory and configmeta.json; we'll no longer re-set the perms each timme configmeta.json is updated.