declare no side effects for the webhooks
Which is indeed true.

The only thing we have which looks like a side effect is this: if a client tries to do a kdcluster/kdconfig status change, any given status change has a one-time-use credentials token as part of it, so if it is a valid token that token is "burned" even by doing a dry run. So why wouldn't we consider that a side effect?

First and most importantly, only KD should be doing status updates (and be able to generate those tokens), and KD won't do dry-run status updates. The conversation really ends there! But, one more point I guess:

Secondly, even if a client did try to re-do the exact same status change after a dry-run attempt, and the change failed, nothing outside of KD token-tracking state is modified and the client would have to be robust to status update failures in any case (and retry with new token).