force blockOwnerDeletion to false in owner refs
closes issue #583

This is to make the OwnerReferencesPermissionEnforcement controller happy -- which is enabled by default in OpenShift, but could also be run elsewhere. And in general it's just a reasonable restriction.

We don't make use of blockOwnerDeletion in any case because we don't use foreground deletion for our resources. But it defaults to true (in the ref returned by NewControllerRef), so since it is causing issues we should explicitly force it to false.