fix incomplete schema areas
issue #611

In the v1 schema regime, the contents of these objects were being discarded.

In the case of the kdstatusbackup, this copy-and-paste of the status schema is not a great solution for long-term development/maintenance (since any future status schema changes now have to be made in two places), but it's a good and easily-verified-as-correct fix for the immediate issue. Will explore other solutions in a following KD release.