slightly different approach to testing for progress data
Previously the code was checking to see if rsync was available here. Technically we should really use generateRsyncInstalledCmd for the test rsync cmd there, if using that approach -- that would make it match up exactly with the initial decision of whether or not to use rsync.

However, instead of making that fix, let's cut to the chase and just check for the existence of the progress file that we're about to try to read.

Also let's make sure to properly interpret the cases where the check itself fails, since this will almost certainly happen for large images while the image is being pulled.

Given that the "check fails" case happens before we can determine whether or not to do rsync, this means that we will populate this field with the "container not ready" message even in the non-rsync case. That being so, then we should also populate it with something after the container does actually come up and start the copy. We'll just say that "no detailed progress info is available" in that case.

(and fix a typo)